fixed Test_multipleCPsAndIterations
Updated the shared round clock to be able to handle progressing
arbitrary amounts of rounds.

Also provided a mechanism to wrap pubsub.PublishSubscibers.

Moved test components to own file.
